chore(core): renamed console hook
