+++ src/utils/useConsole.js
